Simplify chatbot code
