+++ apps/enterprise-knowledge-retrieval/chatbot.py
@@ -1,37 +1,69 @@
-from langchain.agents import Tool
-import pandas as pd
+import os
+
+import openai
 import streamlit as st
-from streamlit_chat import message
-
+from assistant import (answer_question_hyde, answer_user_question, ask_gpt,
+                       initiate_agent)
 from database import get_redis_connection
-from assistant import (
-    answer_user_question,
-    initiate_agent,
-    answer_question_hyde,
-    ask_gpt,
-)
+from langchain.agents import Tool
+from langchain.callbacks import StreamlitCallbackHandler
+from langchain.memory.chat_message_histories import StreamlitChatMessageHistory
+from openai.error import InvalidRequestError
 
-# Initialise database
-
-## Initialise Redis connection
-redis_client = get_redis_connection()
-
-
-### CHATBOT APP
-
-# --- GENERAL SETTINGS ---
-PAGE_TITLE: str = "Knowledge Retrieval Bot"
-PAGE_ICON: str = "🤖"
-
+# General settings
+PAGE_TITLE, PAGE_ICON = "Knowledge Retrieval Bot", "🤖"
 st.set_page_config(page_title=PAGE_TITLE, page_icon=PAGE_ICON)
 
-st.title("Wiki Chatbot")
+st.title("🤖 Wiki Chatbot")
 st.subheader("Learn things - random things!")
 
-# Using object notation
 add_selectbox = st.sidebar.selectbox(
     "What kind of search?", ("Standard vector search", "HyDE")
 )
+
+# Authenticate OpenAI
+if not openai.api_key and "OPENAI_API_KEY" not in os.environ:
+    warning = st.sidebar.info(
+        """
+    No API key provided. You can set your API key in code using 
+    'openai.api_key = <API-KEY>', or you can set the environment 
+    variable `OPENAI_API_KEY=<API-KEY>`. Else, please set your API key below.
+    """
+    )
+    openai.api_key = st.sidebar.text_input("OpenAI API key", type="password")
+    if openai.api_key:
+        warning.empty()
+
+msgs = StreamlitChatMessageHistory()
+token_count = sum(len(str(i)) for i in msgs.messages)
+
+
+def clear_history(token_count):
+    msgs.clear()
+    message = (
+        "How can I help you?"
+        if token_count <= 4097
+        else f"This model's maximum context length is 4097 tokens. Your messages resulted in {token_count} tokens. Resetting chat history."
+    )
+    msgs.add_ai_message(message)
+    st.session_state.steps = {}
+
+
+if not msgs.messages or st.sidebar.button("Reset chat history"):
+    clear_history(token_count)
+
+avatars = {"human": "user", "ai": "assistant"}
+
+for idx, msg in enumerate(msgs.messages):
+    with st.chat_message(avatars[msg.type]):
+        # Render intermediate steps if any were saved
+        for step in st.session_state.steps.get(str(idx), []):
+            if step[0].tool != "_Exception":
+                with st.status(
+                    f"**{step[0].tool}**: {step[0].tool_input}", state="complete"
+                ):
+                    st.write(step[0].log, step[1])
+        st.write(msg.content)
 
 # Define which tools the agent can use to answer user queries
 tools = [
@@ -40,47 +72,26 @@ tools = [
         func=answer_user_question
         if add_selectbox == "Standard vector search"
         else answer_question_hyde,
-        description="Useful for when you need to answer general knowledge questions. Input should be a fully formed question.",
+        description="Answer general knowledge questions.",
     ),
     Tool(
-        name="Ask",
-        func=ask_gpt,
-        description="Useful if the question is not general knowledge. Input should be a fully formed question.",
+        name="Ask", func=ask_gpt, description="Answer non-general knowledge questions."
     ),
 ]
 
-if "generated" not in st.session_state:
-    st.session_state["generated"] = []
-
-if "past" not in st.session_state:
-    st.session_state["past"] = []
-
-
-def query(question):
-    response = st.session_state["chat"].ask_assistant(question)
-    return response
-
-
-prompt = st.text_input("What do you want to know: ", "", key="input")
-
-if st.button("Submit", key="generationSubmit"):
-    with st.spinner("Thinking..."):
-        # Initialization
-        if "agent" not in st.session_state:
-            st.session_state["agent"] = initiate_agent(tools)
-
-        response = st.session_state["agent"].run(prompt)
-
-        st.session_state.past.append(prompt)
-        st.session_state.generated.append(response)
-
-if len(st.session_state["generated"]) > 0:
-    for i in range(len(st.session_state["generated"]) - 1, -1, -1):
-        message(st.session_state["generated"][i], key=str(i))
-        message(st.session_state["past"][i], is_user=True, key=str(i) + "_user")
-
-    with st.expander("See search results"):
-
-        results = list(pd.read_csv("results.csv")["result"])
-
-        st.write(results)
+if prompt := st.chat_input(
+    "Examples: What is light made of? When was Wikipedia founded?"
+):
+    st.chat_message("user").write(prompt)
+    executor = initiate_agent(tools)
+    with st.chat_message("assistant"):
+        st_cb = StreamlitCallbackHandler(st.container())
+        try:
+            response = executor(prompt, callbacks=[st_cb])
+        except InvalidRequestError as e:
+            clear_history(token_count)
+            st.experimental_rerun()
+        st.write(response["output"])
+        st.session_state.steps[str(len(msgs.messages) - 1)] = response[
+            "intermediate_steps"
+        ]
