+++ apps/enterprise-knowledge-retrieval/requirements.txt
@@ -1,4 +1,4 @@
-langchain>=0.0.284
+langchain==0.0.284
 numpy==1.24.2
 openai==0.27.4
 openai[datalib]==0.27.4
